@@ -1,7 +1,7 @@
-import React, { useEffect, useState } from "react";
-import { useHistory, useParams } from "react-router";
-import { listTable, seatingTable } from "../utils/api";
+import { useEffect, useState } from "react";
+import { redirect, useNavigate, useParams } from "react-router";
 import ErrorAlert from "../layout/ErrorAlert";
+import { listTable, seatingTable } from "../utils/api";
 
 const initialFormState = {
 	table_id: "",
@@ -11,7 +11,7 @@ const initialFormState = {
  * Route `/reservations/:reservation_id/seat`
  */
 export default function SetReservationSeat() {
-	const history = useHistory();
+	const navigate = useNavigate();
 	const { reservation_id } = useParams();
 
 	const [formData, setFormData] = useState({ ...initialFormState });
@@ -37,7 +37,7 @@ export default function SetReservationSeat() {
 		seatingTable(formData.table_id, { reservation_id })
 			.then((data) => {
 				console.log(data);
-				history.push("/dashboard");
+				redirect("/dashboard");
 			})
 			.catch(setError);
 	};
@@ -68,7 +68,7 @@ export default function SetReservationSeat() {
 				</label>
 				<br />
 				<div className='btn-group'>
-					<button className='btn btn-dark' onClick={() => history.goBack()}>
+					<button className='btn btn-dark' onClick={() => navigate(-1)}>
 						Cancel
 					</button>
 					<button className='btn btn-dark border-left' type='submit'>
